requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-jsonfield==1.0.1
idna==2.8
PlexAPI==3.1.0
pytz==2018.7
requests==2.21.0
six==1.12.0
tqdm==4.28.1
urllib3==1.24.1
websocket-client==0.54.0
